+++ fractal_server/app/routes/api/v2/job.py
@@ -4,7 +4,6 @@ from pathlib import Path
 
 from fastapi import APIRouter
 from fastapi import Depends
-from fastapi import HTTPException
 from fastapi import Response
 from fastapi import status
 from fastapi.responses import StreamingResponse
@@ -94,42 +93,6 @@ async def get_workflow_jobs(
     )
     job_list = res.scalars().all()
     return job_list
-
-
-@router.get(
-    "/project/{project_id}/latest-job/",
-    response_model=JobRead,
-)
-async def get_latest_job(
-    project_id: int,
-    workflow_id: int,
-    dataset_id: int,
-    user: UserOAuth = Depends(get_api_guest),
-    db: AsyncSession = Depends(get_async_db),
-) -> JobV2:
-    await _get_workflow_check_access(
-        project_id=project_id,
-        workflow_id=workflow_id,
-        user_id=user.id,
-        required_permissions=ProjectPermissions.READ,
-        db=db,
-    )
-    stm = (
-        select(JobV2)
-        .where(JobV2.project_id == project_id)
-        .where(JobV2.workflow_id == workflow_id)
-        .where(JobV2.dataset_id == dataset_id)
-        .order_by(JobV2.start_timestamp.desc())
-        .limit(1)
-    )
-    res = await db.execute(stm)
-    latest_job = res.scalar_one_or_none()
-    if latest_job is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Job with {workflow_id=} and {dataset_id=} not found.",
-        )
-    return latest_job
 
 
 @router.get(