+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * API:
     * Introduce `/star/` and `/unstar/` endpoints for datasets (\#3311).
+    * Remove `/project/{project_id}/status/` and move its logic to `/project/{project_id}/latest-job/` (\#3317).
 * Database:
     * Introduce `DatasetV2.is_starred` (\#3311).
 * Dependencies: